# api/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.InviteUserDTO:
    properties:
      email:
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
      role:
        enum:
        - owner
        - manager
        - staff
        - admin
        type: string
    required:
    - role
    type: object
  dto.JoinTenantDTO:
    properties:
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - name
    - password
    - token
    type: object
...
  dto.Pagination:
    properties:
      currentPage:
        type: integer
      limit:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
...
  dto.UserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      tenantId:
        type: string
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: High performance Go API for AceExtension
  title: AceExtension API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Fetch users with advanced filtering, sorting and pagination
      parameters:
      - description: Search term
        in: query
        name: search
        type: string
      - description: Field to sort by
        in: query
        name: sortBy
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sortOrder
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: JSON encoded filters
        in: query
        name: filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserListResponse'
      security:
      - BearerAuth: []
      summary: List tenant users
      tags:
      - users
  /users/invite:
    post:
      consumes:
      - application/json
      description: Send an invitation to join the tenant
      parameters:
      - description: Invitation Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.InviteUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Invite a new user
      tags:
      - users
  /users/join:
    post:
      consumes:
      - application/json
      description: Accept invitation and create user account
      parameters:
      - description: Join Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.JoinTenantDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Join tenant via invitation
      tags:
      - users
swagger: "2.0"